test playbook should setup ansible structure, PYTHONPATH
Instead of having to do so many path gyrations in the python scripts,
the test playbook should set up the environment on the managed host
to look like the correct Ansible module/plugin structure.  The
playbook knows about the role vs. collection structure and can set
it up correctly for python2 and python3 including the PYTHONPATH.
This also gets rid of the requirement on using `git` and assuming
the test is being run from inside a `git` clone.